+++ app/src/main/java/com/example/pizzapanda/data/PizzaDatabase.kt
@@ -2,8 +2,6 @@ package com.example.pizzapanda.data
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.example.pizzapanda.data.example.ExampleDao
-import com.example.pizzapanda.data.example.ExampleEntity
 import com.example.pizzapanda.data.menu.MenuDao
 import com.example.pizzapanda.data.menu.MenuEntity
 import com.example.pizzapanda.data.order.OrderDao
@@ -13,7 +11,6 @@ import com.example.pizzapanda.data.orderDetails.OrderDetailsEntity
 
 @Database(
     entities = [
-        ExampleEntity::class,
         MenuEntity::class,
         OrderEntity::class,
         OrderDetailsEntity::class
@@ -21,8 +18,6 @@ import com.example.pizzapanda.data.orderDetails.OrderDetailsEntity
     version = 1
 )
 abstract class PizzaDatabase : RoomDatabase() {
-    abstract fun exampleDao(): ExampleDao
-
     abstract fun menuDao(): MenuDao
 
     abstract fun orderDao(): OrderDao
