--- app/src/main/java/com/example/pizzapanda/data/example/ExampleDao.kt
@@ -1,14 +1,0 @@
-package com.example.pizzapanda.data.example
-
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.Query
-
-@Dao
-interface ExampleDao {
-    @Query("SELECT * FROM examples order by id desc")
-    suspend fun getAll(): List<ExampleEntity>
-
-    @Insert
-    suspend fun insert(todo: ExampleEntity)
-}
