--- app/src/main/java/com/example/pizzapanda/data/example/ExampleEntity.kt
@@ -1,14 +1,0 @@
-package com.example.pizzapanda.data.example
-
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "examples")
-data class ExampleEntity(
-    @PrimaryKey(autoGenerate = true)
-    val id: Int? = null,
-
-    @ColumnInfo(name = "name")
-    val name: String,
-)
