--- app/src/main/java/com/example/pizzapanda/domain/model/Example.kt
@@ -1,6 +1,0 @@
-package com.example.pizzapanda.domain.model
-
-data class Example(
-    val id: Int? = null,
-    val name: String = "",
-)
