--- app/src/main/java/com/example/pizzapanda/domain/repository/ExampleRepository.kt
@@ -1,9 +1,0 @@
-package com.example.pizzapanda.domain.repository
-
-import com.example.pizzapanda.domain.model.Example
-
-interface ExampleRepository {
-    suspend fun getExamples(): List<Example>
-
-    suspend fun addExample(example: Example)
-}
