+++ app/src/main/java/com/example/pizzapanda/domain/storage/util/Directory.kt
@@ -1,5 +1,5 @@
 package com.example.pizzapanda.domain.storage.util
 
-sealed class Directory (val path: String){
-    object Image: Directory("images")
+sealed class Directory(val path: String) {
+    object Image : Directory("images")
 }
