--- app/src/main/java/com/example/pizzapanda/domain/usecase/ExampleUseCases.kt
@@ -1,6 +1,0 @@
-package com.example.pizzapanda.domain.usecase
-
-class ExampleUseCases(
-    val getExamples: GetExamplesUseCase,
-    val addExample: AddExampleUseCase
-)
