--- app/src/main/java/com/example/pizzapanda/domain/usecase/GetExamplesUseCase.kt
@@ -1,10 +1,0 @@
-package com.example.pizzapanda.domain.usecase
-
-import com.example.pizzapanda.domain.model.Example
-import com.example.pizzapanda.domain.repository.ExampleRepository
-
-class GetExamplesUseCase(private val exampleRepository: ExampleRepository) {
-    suspend operator fun invoke(): List<Example> {
-        return exampleRepository.getExamples()
-    }
-}
