+++ app/src/main/java/com/example/pizzapanda/presentation/MainActivity.kt
@@ -6,31 +6,21 @@ import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.Button
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.example.pizzapanda.R
-import com.example.pizzapanda.presentation.admin.adminComponents.AdminScreen
-import com.example.pizzapanda.presentation.example.ExampleViewModel
-import com.example.pizzapanda.presentation.example.components.ExampleScreen
-import com.example.pizzapanda.presentation.main.mainComponents.MainScreen
+import com.example.pizzapanda.presentation.admin.components.AdminScreen
+import com.example.pizzapanda.presentation.admin.components.MainScreen
 import com.example.pizzapanda.presentation.util.Screen
 import com.example.pizzapanda.ui.theme.PizzaPandaTheme
 import dagger.hilt.android.AndroidEntryPoint
-import javax.inject.Inject
 
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
-    @Inject
-    lateinit var viewModel: ExampleViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -56,31 +46,21 @@ class MainActivity : ComponentActivity() {
                         }
                         // MSYM Admin Screen
                         composable(route = Screen.AdminScreen.route) {
-                            AdminScreen(viewModel = viewModel) {
+                            AdminScreen {
                                 navController.navigate(Screen.MainScreen.route)
                             }
                         }
                         // MSYM
                         composable(route = Screen.AdminScreen.route) {
-                            AdminScreen(viewModel = viewModel) {
+                            AdminScreen {
                                 navController.navigate(Screen.MainScreen.route)
-                            }
-                        }
-                        composable(route = Screen.HomeScreen.route) {
-                            Greeting("Pizza Panda") {
-                                navController.navigate(Screen.ExampleScreen.route)
-                            }
-                        }
-                        composable(route = Screen.ExampleScreen.route) {
-                            ExampleScreen(viewModel = viewModel) {
-                                navController.navigate(Screen.HomeScreen.route)
                             }
                         }
                         composable(route = Screen.PizzaUserScreen.route) {
                             pizzaJuiceFlag.value = "pizza"
-                            Column() {
-                                Row() {
-                                    Column() {
+                            Column {
+                                Row {
+                                    Column {
                                         MainScreen(goToAdmin = {
                                             navController.navigate(Screen.AdminScreen.route)
                                         })
@@ -90,9 +70,9 @@ class MainActivity : ComponentActivity() {
                         }
                         composable(route = Screen.JuiceUserScreen.route) {
                             pizzaJuiceFlag.value = "juice"
-                            Column() {
-                                Row() {
-                                    Column() {
+                            Column {
+                                Row {
+                                    Column {
                                         MainScreen(goToAdmin = {
                                             navController.navigate(Screen.AdminScreen.route)
                                         })
@@ -103,18 +83,6 @@ class MainActivity : ComponentActivity() {
                     }
                 }
             }
-        }
-    }
-}
-
-@Composable
-fun Greeting(name: String, goToExample: () -> Unit) {
-    Column {
-        Text(text = "Hello $name!")
-        Button(onClick = {
-            goToExample()
-        }) {
-            Text(text = stringResource(id = R.string.show_examples))
         }
     }
 }
