+++ app/src/main/java/com/example/pizzapanda/presentation/admin/components/AdminScreen.kt
@@ -1,0 +1,8 @@
+package com.example.pizzapanda.presentation.admin.components
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun AdminScreen(backToHome: () -> Unit) {
+    AdminOverView(backToHome)
+}
