--- app/src/main/java/com/example/pizzapanda/presentation/example/ExampleEvent.kt
@@ -1,5 +1,0 @@
-package com.example.pizzapanda.presentation.example
-
-sealed class ExampleEvent {
-    data class Add(val name: String) : ExampleEvent()
-}
