--- app/src/main/java/com/example/pizzapanda/presentation/example/ExampleState.kt
@@ -1,5 +1,0 @@
-package com.example.pizzapanda.presentation.example
-
-import com.example.pizzapanda.domain.model.Example
-
-data class ExampleState(val exampleList: List<Example>)
