--- app/src/main/java/com/example/pizzapanda/presentation/main/MainState.kt
@@ -1,4 +1,0 @@
-package com.example.pizzapanda.presentation.main
-
-class MainState {
-}
