+++ app/src/main/java/com/example/pizzapanda/presentation/util/Screen.kt
@@ -1,10 +1,8 @@
 package com.example.pizzapanda.presentation.util
 
 sealed class Screen(val route: String) {
-    object HomeScreen: Screen("home")
-    object ExampleScreen: Screen("example")
-    object MainScreen: Screen("main")
-    object AdminScreen: Screen("admin")
+    object MainScreen : Screen("main")
+    object AdminScreen : Screen("admin")
     object PizzaUserScreen : Screen("pizza")
     object JuiceUserScreen : Screen("juice")
 }
